requirements.txt
asgiref==3.5.0
cycler==0.11.0
Django==4.0.4
django-environ==0.8.1
fonttools==4.33.1
gunicorn==20.1.0
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.8
python-dateutil==2.8.2
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.1.0
tzdata==2022.1
whitenoise==6.0.0
